Performance: double->float and fix cpu utilisation by copying sim state for rendering

+++ physics/damper.h
@@ -6,7 +6,7 @@
 class damper
 {
 public:
-    damper(size_t id_mass1, size_t id_mass2, double k, double l0) : id_mass1_(id_mass1), id_mass2_(id_mass2), k_(k), l0_(l0) {}
+    damper(size_t id_mass1, size_t id_mass2, float k, float l0) : id_mass1_(id_mass1), id_mass2_(id_mass2), k_(k), l0_(l0) {}
 
     damper(size_t id_mass1, size_t id_mass2, const spring& spring) : id_mass1_(id_mass1), id_mass2_(id_mass2)
     {
@@ -17,12 +17,12 @@ public:
     size_t id_mass1() const { return id_mass1_; }
     size_t id_mass2() const { return id_mass2_; }
     
-    double force(double length, double speed) const
+    float force(float length, float speed) const
     {
         return k_* speed * speed;
     }
 
-    double k() const
+    float k() const
     {
         return k_;
     }
@@ -30,7 +30,7 @@ private:
     size_t id_mass1_;
     size_t id_mass2_;
 
-    double k_;      //damping factor
-    double l0_;     //rest length
+    float k_;      //damping factor
+    float l0_;     //rest length
 };
